Simulated Entity Storage
What: Added a simulated entity storage class that does not write to disk

Why: This is useful for testing purposes, as it enables better dependency injection

How: Created a new class `SimEntityStorage` that stores entities in memory.
It uses a map of resource_paths to a pair containing ID and data.

Testing:
- Added unit tests to verify the functionality of the new class.

Change-Id: Icbb1cea9cba2609a20088f6a9cbd2e2ef5efd2a5